# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: number;
  text: string;
  createdAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  text: string;
}

export interface TaskUpdate {
  /** @minLength 1 */
  text: string;
}